+++ jgclark.Dashboard/src/index.js
@@ -1,20 +1,14 @@
 // @flow
-
 // ----------------------------------------------------------------------------
 // Dashboard plugin for NotePlan
 // Jonathan Clark
-// last updated 2.7.2024 for v2.0.0, @jgclark
+// last updated 7.7.2024 for v2.0.1, @jgclark
 // ----------------------------------------------------------------------------
 
 /**
  * Imports
  */
-// allow changes in plugin.json to trigger recompilation
-// import pluginJson from '../plugin.json'
 import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
-// // import { getPluginJson, pluginUpdated, updateSettingData } from '@helpers/NPConfiguration'
-// import { isHTMLWindowOpen, logWindowsList } from '@helpers/NPWindows'
-// import { showMessage } from '@helpers/userInput'
 
 /**
  * Command Exports
@@ -25,7 +19,6 @@ export { editSettings } from '@helpers/NPSettings'
  * Other imports/exports
  */
 
-
 export { decideWhetherToUpdateDashboard } from './dashboardTriggers.js'
 
 export {
@@ -33,7 +26,8 @@ export {
   onMessageFromHTMLView,
   showDemoDashboard,
   setSetting,
+  setSettings,
+  makeSettingsAsCallback,
 } from './reactMain.js'
-
 
 export { onUpdateOrInstall, init, onSettingsUpdated, versionCheck } from './NPHooks'